+++ src/bfd/coff-i860.c
@@ -111,9 +111,9 @@ coff_i860_reloc (bfd *abfd,
 
 	  case 2:
 	    {
-	      long x = bfd_get_32 (abfd, addr);
-	      DOIT (x);
-	      bfd_put_32 (abfd, (bfd_vma) x, addr);
+	      long x = (long)bfd_get_32(abfd, addr);
+	      DOIT(x);
+	      bfd_put_32(abfd, (bfd_vma)x, addr);
 	    }
 	    break;
