+++ src/bfd/coff-x86_64.c
@@ -158,17 +158,17 @@ coff_amd64_reloc (bfd *abfd,
 
 	  case 1:
 	    {
-	      short x = bfd_get_16 (abfd, addr);
-	      DOIT (x);
-	      bfd_put_16 (abfd, (bfd_vma) x, addr);
+	      short x = bfd_get_16(abfd, addr);
+	      DOIT(x);
+	      bfd_put_16(abfd, (bfd_vma)x, addr);
 	    }
 	    break;
 
 	  case 2:
 	    {
-	      long x = bfd_get_32 (abfd, addr);
-	      DOIT (x);
-	      bfd_put_32 (abfd, (bfd_vma) x, addr);
+	      long x = (long)bfd_get_32(abfd, addr);
+	      DOIT(x);
+	      bfd_put_32(abfd, (bfd_vma)x, addr);
 	    }
 	    break;
 	  case 4:
